refactor: extract Dashboard components
