+++ client/src/pages/Dashboard.tsx
@@ -1,65 +1,16 @@
 import { useAuthStore } from '@/stores/authStore';
-import { Calendar, User, Activity, Heart, Sparkles } from 'lucide-react';
+import { Calendar, User, Heart, Sparkles } from 'lucide-react';
 import { useProfile } from '@/hooks/useProfile';
-import { addDays, differenceInDays } from 'date-fns';
 import { useNavigate } from 'react-router';
 import { Button } from '@/components/ui/button';
+import { CycleStats } from '@/features/Dashboard/components/CycleStats';
+import { useCycleCalculations } from '@/features/Dashboard/hooks/useCycleCalculations';
 
 const Dashboard = () => {
   const { user } = useAuthStore();
   const navigate = useNavigate();
   const { data: profile, isLoading } = useProfile();
-
-  // Calculate cycle data
-  const calculateCycleData = () => {
-    if (!profile?.lastPeriodDate) {
-      return {
-        currentDay: '--',
-        nextPeriod: '--',
-        cycleLength: '--',
-      };
-    }
-
-    const lastPeriod = new Date(profile.lastPeriodDate);
-    const today = new Date();
-    const daysSinceLastPeriod = differenceInDays(today, lastPeriod);
-    const cycleLength = profile.cycleLength || 28;
-    const nextPeriodDate = addDays(lastPeriod, cycleLength);
-    const daysUntilNextPeriod = differenceInDays(nextPeriodDate, today);
-
-    return {
-      currentDay: daysSinceLastPeriod + 1,
-      nextPeriod: daysUntilNextPeriod > 0 
-        ? `${daysUntilNextPeriod} days` 
-        : daysUntilNextPeriod === 0 
-        ? 'Today' 
-        : 'Late',
-      cycleLength: `${cycleLength} days`,
-    };
-  };
-
-  const cycleData = calculateCycleData();
-
-  const stats = [
-    {
-      title: 'Current Cycle Day',
-      value: cycleData.currentDay,
-      icon: Activity,
-      description: profile ? 'Days since last period' : 'Set up your profile to track cycles',
-    },
-    {
-      title: 'Next Period',
-      value: cycleData.nextPeriod,
-      icon: Calendar,
-      description: profile?.lastPeriodDate ? 'Estimated start date' : 'Add your last period date',
-    },
-    {
-      title: 'Average Cycle',
-      value: cycleData.cycleLength,
-      icon: User,
-      description: profile ? 'Your cycle length' : 'Complete profile setup',
-    },
-  ];
+  const cycleData = useCycleCalculations(profile);
 
   return (
     <div className="min-h-[calc(100vh-4rem)]">
@@ -74,34 +25,7 @@ const Dashboard = () => {
       </div>
 
       {/* Stats Grid with Animation */}
-      <div className="grid grid-cols-1 md:grid-cols-3 gap-6 mb-10 px-6">
-        {stats.map((stat, index) => {
-          const Icon = stat.icon;
-          return (
-            <div
-              key={stat.title}
-              className="group relative bg-white rounded-2xl shadow-md hover:shadow-xl border border-muted/20 p-6 transition-all duration-300 hover:-translate-y-1"
-              style={{ animationDelay: `${index * 100}ms` }}
-            >
-              <div className="absolute inset-0 bg-gradient-to-br from-primary/5 to-transparent rounded-2xl opacity-0 group-hover:opacity-100 transition-opacity" />
-              <div className="relative flex items-center justify-between">
-                <div className="flex-1">
-                  <p className="text-sm font-medium text-muted-foreground">{stat.title}</p>
-                  <p className="text-3xl font-bold text-foreground mt-2 bg-gradient-to-r from-foreground to-foreground/80 bg-clip-text">
-                    {stat.value}
-                  </p>
-                  <p className="text-xs text-muted-foreground mt-2 leading-relaxed">
-                    {stat.description}
-                  </p>
-                </div>
-                <div className="p-4 bg-gradient-to-br from-primary/20 to-primary/10 rounded-2xl group-hover:scale-110 transition-transform">
-                  <Icon className="h-6 w-6 text-primary" />
-                </div>
-              </div>
-            </div>
-          );
-        })}
-      </div>
+      <CycleStats profile={profile} isLoading={isLoading} />
 
       {/* Cycle Progress Indicator */}
       {profile && (
